+++ plugins/redcar/redcar.rb
@@ -86,7 +86,7 @@ module Redcar
           tab = win.new_tab(Redcar::EditTab)
         else
           window = Redcar.app.new_window
-          tab = window.new_tab(Redcar::EditTab)          
+          tab = window.new_tab(Redcar::EditTab)
         end
         tab.title = "untitled"
         tab.focus
@@ -221,9 +221,25 @@ module Redcar
 
     class CloseTreeCommand < Command
       def execute
-        treebook = Redcar.app.focussed_window.treebook
-        tree = treebook.focussed_tree
-        treebook.remove_tree(tree)
+        win = Redcar.app.focussed_window
+        if win and treebook = win.treebook
+          if tree = treebook.focussed_tree
+            treebook.remove_tree(tree)
+          end
+        end
+      end
+    end
+
+    class ToggleTreesCommand < Command
+      def execute
+        win = Redcar.app.focussed_window
+        if win and treebook = win.treebook
+          if win.trees_visible?
+            win.set_trees_visible(false)
+          else
+            win.set_trees_visible(true)
+          end
+        end
       end
     end
 
@@ -423,7 +439,7 @@ Redcar.environment: #{Redcar.environment}
         end
         doc.ensure_visible(doc.cursor_offset)
         doc.insert(doc.cursor_offset, "\n")
-        
+
       end
     end
 
@@ -569,30 +585,34 @@ Redcar.environment: #{Redcar.environment}
         cursor_line  = doc.cursor_line
         cursor_line_offset = doc.cursor_line_offset
         diff = 0
-        doc.controllers(AutoIndenter::DocumentController).first.disable do
-          doc.selection_ranges.each do |range|
-            doc.delete(range.begin - diff, range.count)
-            diff += range.count
-          end
-          texts = Redcar.app.clipboard.last.dup
-          texts.each_with_index do |text, i|
-            line_ix = start_line + i
-            if line_ix == doc.line_count
-              doc.insert(doc.length, "\n" + " "*line_offset)
-            else
-              line = doc.get_line(line_ix).chomp
-              if line.length < line_offset
-                doc.insert(
-                doc.offset_at_inner_end_of_line(line_ix),
-                " "*(line_offset - line.length)
-                )
-              end
+        doc.controllers(AutoPairer::DocumentController).first.disable do
+          doc.controllers(AutoIndenter::DocumentController).first.disable do
+            doc.controllers(AutoCompleter::DocumentController).first.start_modification
+            doc.selection_ranges.each do |range|
+              doc.delete(range.begin - diff, range.count)
+              diff += range.count
             end
-            doc.insert(
-            doc.offset_at_line(line_ix) + line_offset,
-            text
-            )
-            doc.cursor_offset = doc.offset_at_line(line_ix) + line_offset + text.length
+            texts = Redcar.app.clipboard.last.dup
+            texts.each_with_index do |text, i|
+              line_ix = start_line + i
+              if line_ix == doc.line_count
+                doc.insert(doc.length, "\n" + " "*line_offset)
+              else
+                line = doc.get_line(line_ix).chomp
+                if line.length < line_offset
+                  doc.insert(
+                  doc.offset_at_inner_end_of_line(line_ix),
+                  " "*(line_offset - line.length)
+                  )
+                end
+              end
+              doc.insert(
+              doc.offset_at_line(line_ix) + line_offset,
+              text
+              )
+              doc.cursor_offset = doc.offset_at_line(line_ix) + line_offset + text.length
+            end
+            doc.controllers(AutoCompleter::DocumentController).first.end_modification
           end
         end
       end
@@ -626,15 +646,15 @@ Redcar.environment: #{Redcar.environment}
         cursor_ix = doc.cursor_offset
         if doc.selection?
           start_ix = doc.selection_range.begin
-          text = doc.selected_text                              
-    
+          text = doc.selected_text
+
           sorted_text = text.split("\n").sort().join("\n")
           doc.replace_selection(sorted_text)
           doc.cursor_offset = cursor_ix
         end
       end
     end
-    
+
     class DialogExample < Redcar::Command
       def execute
       	builder = Menu::Builder.new do
@@ -660,23 +680,22 @@ Redcar.environment: #{Redcar.environment}
 
         button :go, "Go", "Return" do
           new_line_ix = line.value.to_i - 1
-          if new_line_ix < doc.line_count and new_line_ix >= 0
-            doc.cursor_offset = doc.offset_at_line(new_line_ix)
-            doc.scroll_to_line(new_line_ix)
-            win.close_speedbar
+          if new_line_ix < @doc.line_count and new_line_ix >= 0
+            @doc.cursor_offset = @doc.offset_at_line(new_line_ix)
+            @doc.scroll_to_line(new_line_ix)
+            @win.close_speedbar
           end
         end
 
-        def initialize(command)
+        def initialize(command, win)
           @command = command
+          @doc     = command.doc
+          @win     = win
         end
-
-        def doc; @command.doc; end
-        def win; @command.send(:win); end
       end
 
       def execute
-        @speedbar = GotoLineCommand::Speedbar.new(self)
+        @speedbar = GotoLineCommand::Speedbar.new(self, win)
         win.open_speedbar(@speedbar)
       end
     end
@@ -697,6 +716,12 @@ Redcar.environment: #{Redcar.environment}
           notebook = Redcar.app.focussed_window_notebook
           notebook.tabs[tab_num-1].focus if notebook.tabs[tab_num-1]
         end
+      end
+    end
+
+    class ToggleFullscreen < Command
+      def execute
+        Redcar.app.focussed_window.fullscreen = !Redcar.app.focussed_window.fullscreen
       end
     end
 
@@ -762,11 +787,10 @@ Redcar.environment: #{Redcar.environment}
         #link "Cmd+Ctrl+O",  Project::OpenRemoteCommand
         link "Cmd+S",       Project::FileSaveCommand
         link "Cmd+Shift+S", Project::FileSaveAsCommand
-        link "Cmd+Ctrl+R",  Project::RevealInProjectCommand
         link "Cmd+W",       CloseTabCommand
         link "Cmd+Shift+W", CloseWindowCommand
         link "Cmd+Q",       QuitCommand
-        
+
         #link "Cmd+Return",   MoveNextLineCommand
 
         link "Cmd+Shift+E", EditView::InfoSpeedbarCommand
@@ -787,6 +811,7 @@ Redcar.environment: #{Redcar.environment}
         link "Cmd+Shift+I", AutoIndenter::IndentCommand
         link "Cmd+L",       GotoLineCommand
         link "Cmd+F",       DocumentSearch::SearchForwardCommand
+        link "Cmd+H",       DocumentSearch::SearchAndReplaceCommand
         #link "Cmd+Shift+F", DocumentSearch::RepeatPreviousSearchForwardCommand
         link "Cmd+Ctrl+F",  DocumentSearch::SearchAndReplaceCommand
         link "Cmd+Shift+F", Redcar::FindInProject::OpenSearch
@@ -810,11 +835,13 @@ Redcar.environment: #{Redcar.environment}
         link "Cmd+Shift+]",     SwitchTabUpCommand
         link "Ctrl+Shift+[",    MoveTabDownCommand
         link "Ctrl+Shift+]",    MoveTabUpCommand
+        link "F11",             ToggleFullscreen
         link "Cmd+Alt+I",       ToggleInvisibles
         link "Ctrl+R",          Runnables::RunEditTabCommand
         link "Cmd+I",           OutlineView::OpenOutlineViewCommand
 
         link "Ctrl+Shift+P",    PrintScopeCommand
+        link "Cmd+Shift+H",     ToggleTreesCommand
 
         link "Cmd+Shift+R",     PluginManagerUi::ReloadLastReloadedCommand
 
@@ -837,11 +864,10 @@ Redcar.environment: #{Redcar.environment}
         #link "Alt+Shift+O",  Project::OpenRemoteCommand
         link "Ctrl+S",       Project::FileSaveCommand
         link "Ctrl+Shift+S", Project::FileSaveAsCommand
-        link "Ctrl+Shift+R", Project::RevealInProjectCommand
         link "Ctrl+W",       CloseTabCommand
         link "Ctrl+Shift+W", CloseWindowCommand
         link "Ctrl+Q",       QuitCommand
-        
+
         link "Ctrl+Enter",   MoveNextLineCommand
 
         link "Ctrl+Shift+E", EditView::InfoSpeedbarCommand
@@ -862,6 +888,7 @@ Redcar.environment: #{Redcar.environment}
         link "Ctrl+Shift+[", AutoIndenter::IndentCommand
         link "Ctrl+L",       GotoLineCommand
         link "Ctrl+F",       DocumentSearch::SearchForwardCommand
+        link "Ctrl+H",       DocumentSearch::SearchAndReplaceCommand
         link "F3",           DocumentSearch::RepeatPreviousSearchForwardCommand
         link "Ctrl+Shift+F", Redcar::FindInProject::OpenSearch
         link "Ctrl+A",       SelectAllCommand
@@ -885,16 +912,18 @@ Redcar.environment: #{Redcar.environment}
         link "Ctrl+Shift+P",    PrintScopeCommand
 
         link "Ctrl+Alt+O",       SwitchNotebookCommand
-
+        link "Ctrl+Shift+H",     ToggleTreesCommand
         link "Ctrl+Page Up",         SwitchTabDownCommand
         link "Ctrl+Page Down",       SwitchTabUpCommand
         link "Ctrl+Shift+Page Up",   MoveTabDownCommand
         link "Ctrl+Shift+Page Down", MoveTabUpCommand
         link "Ctrl+Shift+R",     PluginManagerUi::ReloadLastReloadedCommand
+        link "F11",              ToggleFullscreen
         link "Ctrl+Alt+I",       ToggleInvisibles
         link "Ctrl+I",           OutlineView::OpenOutlineViewCommand
 
         link "Ctrl+Alt+S", Snippets::OpenSnippetExplorer
+
         #Textmate.attach_keybindings(self, :linux)
 
         # map SelectTab<number>Command
@@ -919,10 +948,9 @@ Redcar.environment: #{Redcar.environment}
         item "New Notebook", :command => NewNotebookCommand, :icon => File.join(Redcar::ICONS_DIRECTORY, "book--plus.png"), :barname => :edit
         item "Close Notebook", :command => CloseNotebookCommand, :icon => File.join(Redcar::ICONS_DIRECTORY, "book--minus.png"), :barname => :edit
       end
-
     end
 
-    def self.menus
+    def self.menus(window)
       Menu::Builder.build do
         sub_menu "File", :priority => :first do
           group(:priority => :first) do
@@ -972,7 +1000,7 @@ Redcar.environment: #{Redcar.environment}
               item "Toggle Block Selection", ToggleBlockSelectionCommand
             end
           end
-          
+
           group(:priority => 40) do
             sub_menu "Document Navigation" do
               item "Goto Line", GotoLineCommand
@@ -1005,6 +1033,8 @@ Redcar.environment: #{Redcar.environment}
             item "Font Size", SelectFontSize
             item "Theme", SelectTheme
           end
+          item "Toggle Tree Visibility", :command => ToggleTreesCommand
+          item "Toggle Fullscreen", :command => ToggleFullscreen, :type => :check, :active => window ? window.fullscreen : false
           separator
           item "New Notebook", NewNotebookCommand
           item "Close Notebook", CloseNotebookCommand
@@ -1038,7 +1068,7 @@ Redcar.environment: #{Redcar.environment}
           end
         end
         sub_menu "Help", :priority => :last do
-          group(:priority => :first) do
+          group(:priority => :last) do
             item "About", AboutCommand
             item "New In This Version", ChangelogCommand
           end
@@ -1093,7 +1123,8 @@ Redcar.environment: #{Redcar.environment}
         s = Time.now
         Redcar::Project::Manager.start(args)
         puts "project start took #{Time.now - s}s"
-        Redcar.app.make_sure_at_least_one_window_open
+        win = Redcar.app.make_sure_at_least_one_window_open
+        win.close if win and args.include?("--no-window")
       end
       Redcar.update_gui do
         Swt.splash_screen.close if Swt.splash_screen
